Magic :: Lesson 14 :: Items casting spells :: WIP
